fix: 레포지토리 테스트 데이터 임시 추가 및 심박수 그래프 평균 심박수 데이터 formatException 관련 수정

+++ app/src/main/java/com/san/heartratemonitormobile/view/screen/ReportDetailActivity.kt
@@ -123,7 +123,7 @@ class ReportDetailActivity : AppCompatActivity() {
         val dataset = arrayListOf<ILineDataSet>(set)
         val data = LineData(dataset)
         binding.chartDayHeartRate.data = data
-        binding.txtAvgHeartRate.text = String.format(HEART_RATE_MESSAGE, viewModel.heartRateData.average())
+        binding.txtAvgHeartRate.text = String.format(HEART_RATE_MESSAGE, viewModel.heartRateData.average().toInt())
         binding.txtMaxHeartRate.text = String.format(HEART_RATE_MESSAGE, viewModel.heartRateData.max())
     }
